removes old examples, updates cli, removes unused files
